pyproject.toml
[tool.poetry]
name = "cnter"
version = "0.1.0"
description = "counter-based social app"
authors = ["sverdejot"]

[tool.poetry.dependencies]
python = "^3.10"
SQLAlchemy = "^1.4.40"
python-dotenv = "^0.20.0"
mariadb = "^1.1.3"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
